Feature#106

+++ server/src/controllers/reservation/ReservationId.ts
@@ -1,25 +1,43 @@
 import { Request, Response } from 'express';
 import { getRepository } from 'typeorm';
 import { reservation } from '../../entity/reservation';
+import { users } from '../../entity/users';
+import { authorizeToken } from '../jwt/AuthorizeToken';
 
 export default {
   patch: async (req: Request, res: Response) => {
-    const reservationId: number = Number(req.params.reservationId);
-    const reservation_status: number = req.body.reservation_status;
+    const decoded = await authorizeToken(req, res);
+    const userRepository = await getRepository(users);
+    const user = await userRepository.findOne({ email: decoded.email });
 
-    const reservationRepository = getRepository(reservation);
-    await reservationRepository.update(reservationId, { reservation_status });
+    if (!user) {
+      res.status(401).json({ message: 'Unauthorized User' });
+    } else {
+      const reservationId: number = Number(req.params.reservationId);
+      const reservation_status: number = req.body.reservation_status;
 
-    res
-      .status(200)
-      .json({ message: 'Reservation Status Successfully Updated' });
+      const reservationRepository = getRepository(reservation);
+      await reservationRepository.update(reservationId, { reservation_status });
+
+      res
+        .status(200)
+        .json({ message: 'Reservation Status Successfully Updated' });
+    }
   },
   delete: async (req: Request, res: Response) => {
-    const reservationId: number = Number(req.params.reservationId);
+    const decoded = await authorizeToken(req, res);
+    const userRepository = await getRepository(users);
+    const user = await userRepository.findOne({ email: decoded.email });
 
-    const reservationRepository = getRepository(reservation);
-    await reservationRepository.delete(reservationId);
+    if (!user) {
+      res.status(401).json({ message: 'Unauthorized User' });
+    } else {
+      const reservationId: number = Number(req.params.reservationId);
 
-    res.status(200).json({ message: 'Reservation Successfully Canceled' });
+      const reservationRepository = getRepository(reservation);
+      await reservationRepository.delete(reservationId);
+
+      res.status(200).json({ message: 'Reservation Successfully Canceled' });
+    }
   },
 };
